Moving function "application_files" to utilities.py in order to prevent circular import. Also updating factory.active_fib_applications to use the application_files function.
